Moved Poly to separate class Polynomial
